backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.6.0
python-dotenv==0.19.0
requests==2.26.0
python-multipart==0.0.6
pymongo